requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
Django==2.1
django-bootstrap3==10.0.1
django-generic-json-views==0.8
django-mathfilters==0.4.0
django-search-views==0.3.7
idna==2.7
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
six==1.10.0
